cloud_functions/lstm_vertexai_predictions/requirements.txt
functions-framework==3.4.0
psycopg2-binary==2.9.7
numpy
pandas
scikit-learn
tensorflow
google-cloud-aiplatform
